Migrate output format options to mft_formatter.py
